CI/tasks/requirements.txt
pyyaml
avionix
dagger-io